+++ libs/insight-viewer/src/Viewer/LineViewer/index.tsx
@@ -27,6 +27,7 @@ export function LineViewer({
     <>
       {showOutline && (
         <polyline
+          className="annotation-line pointer"
           style={{
             ...svgWrapperStyle[isHoveredAnnotation ? 'hoveredOutline' : 'outline'],
             ...polygonAttributes?.style,
@@ -39,6 +40,7 @@ export function LineViewer({
         <>
           {showOutline && (
             <polyline
+              className="annotation-line arrow pointer"
               style={{
                 ...svgWrapperStyle[isHoveredAnnotation ? 'hoveredOutline' : 'outline'],
                 ...polygonAttributes?.style,
@@ -46,10 +48,15 @@ export function LineViewer({
               points={headPoints}
             />
           )}
-          <polyline style={{ ...svgWrapperStyle.default, ...polygonAttributes?.style }} points={headPoints} />
+          <polyline
+            className="annotation-line arrow pointer"
+            style={{ ...svgWrapperStyle.default, ...polygonAttributes?.style }}
+            points={headPoints}
+          />
         </>
       )}
       <polyline
+        className="annotation-line pointer"
         style={{
           ...svgWrapperStyle.default,
           ...polygonAttributes?.style,
@@ -58,6 +65,7 @@ export function LineViewer({
         points={polygonPoints}
       />
       <polyline
+        className="annotation-line pointer"
         style={{
           ...svgWrapperStyle.extendsArea,
           ...polygonAttributes?.style,
@@ -66,7 +74,12 @@ export function LineViewer({
         points={polygonPoints}
       />
       {showAnnotationLabel && labelPosition && (
-        <text style={{ ...textStyle.default }} x={labelPosition[0]} y={labelPosition[1]}>
+        <text
+          className="annotation-line label pointer"
+          style={{ ...textStyle.default }}
+          x={labelPosition[0]}
+          y={labelPosition[1]}
+        >
           {polygonLabel}
         </text>
       )}