+++ libs/insight-viewer/src/Viewer/PolylineDrawer/index.tsx
@@ -48,40 +48,50 @@ export function PolylineDrawer({
   }
 
   const labelPosition = selectedAnnotationLabel ? polylabel([points.map(pixelToCanvas)]) : null
+  const isDrawing = isPolygonSelected ? undefined : 'isDrawing'
 
   return (
     <g data-cy-annotation onMouseDown={() => setAnnotationEditMode('move')}>
       {points && points.length > 0 && (
         <>
-          <PolylineElement isPolygon={!!isPolygonSelected} style={svgWrapperStyle.outline} points={polylinePoints} />
+          <PolylineElement isPolygon={isPolygonSelected} style={svgWrapperStyle.outline} points={polylinePoints} />
           {lineHead === 'arrow' && (
             <>
               <PolylineElement
-                isPolygon={!!isPolygonSelected}
+                className={`annotation-polyline pointer ${isDrawing}`}
+                isPolygon={isPolygonSelected}
                 style={svgWrapperStyle.outline}
                 points={getArrowPoints()}
               />
               <PolylineElement
-                isPolygon={!!isPolygonSelected}
+                className={`annotation-polyline pointer ${isDrawing}`}
+                isPolygon={isPolygonSelected}
                 style={svgWrapperStyle[isSelectedMode ? 'select' : 'default']}
                 points={getArrowPoints()}
               />
             </>
           )}
           <PolylineElement
-            isPolygon={!!isPolygonSelected}
+            className={`annotation-polyline pointer ${isDrawing}`}
+            isPolygon={isPolygonSelected}
             style={svgWrapperStyle[isSelectedMode ? 'select' : 'default']}
             points={polylinePoints}
           />
           <PolylineElement
-            isPolygon={!!isPolygonSelected}
+            className={`annotation-polyline pointer ${isDrawing}`}
+            isPolygon={isPolygonSelected}
             style={svgWrapperStyle.extendsArea}
             points={polylinePoints}
           />
         </>
       )}
       {showAnnotationLabel && selectedAnnotationLabel && labelPosition && (
-        <text style={{ ...textStyle.default }} x={labelPosition[0]} y={labelPosition[1]}>
+        <text
+          className={`annotation-polyline label pointer ${isDrawing}`}
+          style={{ ...textStyle.default }}
+          x={labelPosition[0]}
+          y={labelPosition[1]}
+        >
           {selectedAnnotationLabel}
         </text>
       )}